chore: refactor formatPrice function
