+++ src/utilities/formatPrice.jsx
@@ -1,0 +1,8 @@
+const formatPrice = (price) => {
+  return new Intl.NumberFormat("en-US", {
+    minimumFractionDigits: 0,
+    maximumFractionDigits: 0,
+  }).format(price);
+};
+
+export default formatPrice;
